+++ history.md
@@ -2,3 +2,7 @@
 
 * Fix restart of tsserver not working
 * Fix edit of current buffer change jumplist by using `:keepjumps`
+
+# 2018-07-13
+
+* **Break change:** `diagnoctic` in setting.json changed to `diagnostic`